legacy/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
eventlet>=0.35.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
